# backend/app/models/student.py
from datetime import datetime
from uuid import uuid4

from sqlalchemy import DateTime, ForeignKey, Integer, String
from sqlalchemy.dialects.postgresql import JSONB
from sqlalchemy.orm import Mapped, mapped_column, relationship

from app.db import Base


class Student(Base):
    __tablename__ = "students"

    id: Mapped[str] = mapped_column(String, primary_key=True, default=lambda: str(uuid4()))
    teacher_id: Mapped[str] = mapped_column(String, nullable=False, index=True)
    class_id: Mapped[str] = mapped_column(String, index=True)
    # Only a salted hash lives in PostgreSQL; names live in per-teacher files.
    display_name: Mapped[str] = mapped_column(String)
    grade: Mapped[int] = mapped_column(Integer)
    created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)

    profile: Mapped["StudentProfile"] = relationship(
        back_populates="student", uselist=False, cascade="all, delete-orphan"
    )


class StudentProfile(Base):
    __tablename__ = "student_profiles"

    student_id: Mapped[str] = mapped_column(
        String, ForeignKey("students.id"), primary_key=True
    )
    submissions_count: Mapped[int] = mapped_column(Integer, default=0)
    avg_score: Mapped[float] = mapped_column(default=0.0)
    memory: Mapped[dict] = mapped_column(JSONB, default=dict)
    updated_at: Mapped[datetime] = mapped_column(
        DateTime, default=datetime.utcnow, onupdate=datetime.utcnow
    )

    student: Mapped[Student] = relationship(back_populates="profile")