# openapi.yaml
openapi: 3.0.3
info:
  title: Помощник учителя API
  version: 0.2.0
  description: |
    Backend API для чат-бота и мини-приложения проверки контрольных
    работ в MAX.

    Все ресурсы принадлежат учителю. Аутентификация:

    - MiniApp: заголовок `X-Init-Data` c подписанным MAX WebApp initData.
      Backend валидирует HMAC по `MAX_BOT_TOKEN` и извлекает user_id.
    - Бот (service-to-service): `Authorization: Bearer <MAX_BOT_TOKEN>` +
      `X-Teacher-Id: <user_id>`. Bearer сравнивается с ботовым токеном
      в constant-time, teacher_id доверяется из заголовка. Не выставлять
      этот путь наружу — только внутри compose-сети.

    Кросс-тенантный доступ запрещён: попытка чтения/изменения
    чужого ресурса → 403.
servers:
  - url: /api
    description: Через nginx proxy мини-приложения
  - url: http://localhost:8000
    description: Прямое подключение к backend
security:
  - MaxInitData: []
  - BotService: []

paths:
  /health:
    get:
      summary: Проверка живости сервиса
      security: []
      responses:
        "200":
          description: OK

  /works:
    post:
      summary: Создать шаблон работы (тема, задания, эталон)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkTemplateCreate"
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkTemplate"
    get:
      summary: Список шаблонов работ учителя
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1 }
        - in: query
          name: offset
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorkTemplate" }

  /works/generate:
    post:
      summary: Сгенерировать и сохранить контрольную работу
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateWorkRequest" }
      responses:
        "201":
          description: Работа создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkTemplate" }
        "401":
          description: Неверный или просроченный initData
        "422":
          description: Некорректные параметры генерации

  /works/{work_id}:
    get:
      summary: Шаблон работы учителя
      parameters:
        - in: path
          name: work_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkTemplate" }

  /submissions:
    post:
      summary: Загрузить фото работы одного ученика
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [work_id, student_id, photos]
              properties:
                work_id: { type: string }
                student_id: { type: string }
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "202":
          description: Принято на проверку
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Submission" }
        "403": { $ref: "#/components/responses/Forbidden" }

    get:
      summary: Очередь работ учителя
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, checked, confirmed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Submission" }

  /submissions/{submission_id}:
    get:
      summary: Результат проверки одной работы с разметкой
      parameters:
        - in: path
          name: submission_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  /submissions/{submission_id}/review:
    patch:
      summary: Учитель подтверждает или правит вердикт по заданиям
      parameters:
        - in: path
          name: submission_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeacherReview" }
      responses:
        "200":
          description: OK
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  /students:
    post:
      summary: Создать ученика в классе учителя
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentCreate" }
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
    get:
      summary: Список учеников учителя
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Student" }

  /students/{student_id}/profile:
    get:
      summary: Долгосрочный профиль ученика (память)
      parameters:
        - in: path
          name: student_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentProfile" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  /students/{student_id}:
    delete:
      summary: Удалить ученика и его работы
      parameters:
        - in: path
          name: student_id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Удалено

  /classes/{class_id}/dashboard:
    get:
      summary: Агрегированная статистика по классу учителя
      parameters:
        - in: path
          name: class_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClassDashboard" }
        "403": { $ref: "#/components/responses/Forbidden" }

  /teachers/me/state:
    get:
      summary: Текущий выбор работы и ученика
      description: |
        Общий стейт для бота и мини-приложения. Что учитель выбрал в
        MiniApp — бот увидит здесь, и наоборот. Возвращает пустой
        стейт при первом запросе (200, поля null).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherState" }
    put:
      summary: Установить работу и/или ученика
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_work_id: { type: string, nullable: true }
                current_student_id: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherState" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  /storage/{path}:
    get:
      summary: Скачать файл работы (фото)
      description: |
        StaticFiles-mount, отдающий содержимое каталога submissions
        по относительным путям, которые возвращаются в
        `Submission.photos`. Требует MAX initData.
      parameters:
        - in: path
          name: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Бинарные данные
          content:
            application/octet-stream:
              schema: { type: string, format: binary }

components:
  securitySchemes:
    MaxInitData:
      type: apiKey
      in: header
      name: X-Init-Data
    BotService:
      type: http
      scheme: bearer
      description: |
        Only usable service-to-service inside the compose network.
        Bearer value = MAX_BOT_TOKEN. Additionally requires
        `X-Teacher-Id: <max_user_id>` header.

  responses:
    Forbidden:
      description: Ресурс принадлежит другому учителю
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }

  schemas:
    TeacherState:
      type: object
      properties:
        teacher_id: { type: string }
        current_work_id: { type: string, nullable: true }
        current_student_id: { type: string, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }

    GenerateWorkRequest:
      type: object
      required: [topic, grade, n_tasks]
      properties:
        topic:
          type: string
          minLength: 1
          maxLength: 120
        grade:
          type: integer
          enum: [2, 3, 4]
        n_tasks:
          type: integer
          minimum: 1
          maximum: 20

    StudentCreate:
      type: object
      required: [class_id, display_name, grade]
      properties:
        class_id: { type: string }
        display_name: { type: string }
        grade: { type: integer }

    Student:
      allOf:
        - $ref: "#/components/schemas/StudentCreate"
        - type: object
          properties:
            id: { type: string }
            teacher_id: { type: string }
            created_at: { type: string, format: date-time }

    WorkTemplateCreate:
      type: object
      required: [title, grade, tasks]
      properties:
        title: { type: string, example: "Контрольная №3. Умножение" }
        subject: { type: string, example: "math" }
        grade: { type: integer, example: 3 }
        tasks:
          type: array
          items: { $ref: "#/components/schemas/TaskDefinition" }

    WorkTemplate:
      allOf:
        - $ref: "#/components/schemas/WorkTemplateCreate"
        - type: object
          properties:
            id: { type: string }
            teacher_id: { type: string }
            created_at: { type: string, format: date-time }

    TaskDefinition:
      type: object
      required: [index, statement, expected_answer]
      properties:
        index: { type: integer }
        statement: { type: string }
        expected_answer: { type: string }
        max_points: { type: number, default: 1 }

    Submission:
      type: object
      properties:
        id: { type: string }
        work_id: { type: string }
        student_id: { type: string }
        status:
          type: string
          enum: [pending, checked, confirmed]
        created_at: { type: string, format: date-time }
        photos:
          type: array
          items: { type: string, description: "относительный URL" }

    SubmissionResult:
      allOf:
        - $ref: "#/components/schemas/Submission"
        - type: object
          properties:
            per_task:
              type: array
              items: { $ref: "#/components/schemas/TaskCheck" }
            total_score: { type: number }
            confidence: { type: number, description: "средняя уверенность модели, 0..1" }

    TaskCheck:
      type: object
      properties:
        task_index: { type: integer }
        student_answer: { type: string }
        expected_answer: { type: string }
        is_correct: { type: boolean }
        confidence: { type: number }
        explanation: { type: string, description: "объяснение ошибки для ученика" }
        error_type:
          type: string
          nullable: true
          enum: [вычислительная, методологическая, невнимательность, "не распознано"]
        reasoning_graph:
          type: array
          description: "шаги решения ученика (если восстановлены)"
          items:
            type: object
            properties:
              step: { type: string }
              ok: { type: boolean }
        photo_boxes:
          type: array
          description: "координаты области с ответом на фото"
          items:
            type: object
            properties:
              photo_index: { type: integer }
              x: { type: number }
              y: { type: number }
              w: { type: number }
              h: { type: number }
        teacher_verdict:
          type: object
          nullable: true
          properties:
            is_correct: { type: boolean }
            comment: { type: string }

    TeacherReview:
      type: object
      properties:
        per_task:
          type: array
          items:
            type: object
            required: [task_index, is_correct]
            properties:
              task_index: { type: integer }
              is_correct: { type: boolean }
              comment: { type: string }

    StudentProfile:
      type: object
      properties:
        student_id: { type: string }
        submissions_count: { type: integer }
        avg_score: { type: number }
        weak_topics:
          type: array
          items:
            type: object
            properties:
              topic: { type: string }
              error_rate: { type: number }
        recurring_mistakes:
          type: array
          items: { type: string }
        trend:
          type: string
          enum: [improving, stable, regressing]

    ClassDashboard:
      type: object
      properties:
        class_id: { type: string }
        students_count: { type: integer }
        avg_score: { type: number }
        weak_topics:
          type: array
          items:
            type: object
            properties:
              topic: { type: string }
              error_rate: { type: number }
        students_needing_help:
          type: array
          items:
            type: object
            properties:
              student_id: { type: string }
              reason: { type: string }